split player/enemy tanks.
